Copy a test method to its correct place
